qcacld-3.0: Populate OCV capability for LFR3.0 roaming
In OCV WFA 15.5.1 case, SAE+FT roaming is triggered, as it's LFR3
roaming, host should set ocv capability in RSO config if OCV
capable, else FW will not carry OCI element in Reassoc frame and
thus fail to pass WFA 15.5.1 case.

CRs-Fixed: 3106190

+++ components/umac/mlme/connection_mgr/dispatcher/inc/wlan_cm_roam_public_struct.h
@@ -254,7 +254,7 @@ struct wlan_chan_list {
  * @roam_scan_freq_lst: roam freq list
  */
 struct rso_config {
-	uint8_t rsn_cap;
+	uint16_t rsn_cap;
 	bool disable_hi_rssi;
 	bool roam_control_enable;
 	uint8_t rescan_rssi_delta;
